+++ compiler/src/utils.mli
@@ -180,6 +180,7 @@ type warning =
   | SimplifyVectorSuffix
   | DuplicateVar 
   | UnusedVar 
+  | SplitMemoryAccess
   | SCTchecker
   | Deprecated
   | Experimental